// src/types/common/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  applicants: {
    Tables: {
// ...
      applicant_requirement_media: {
        Row: {
          alt_text: string | null
          applicant_requirement_id: number | null
          created_at: string
          description: string | null
          filename: string | null
          id: number
          mime_type: string | null
          path: string | null
          profiles_id: string | null
          size: number | null
          updated_at: string | null
        }
        Insert: {
          alt_text?: string | null
          applicant_requirement_id?: number | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profiles_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Update: {
          alt_text?: string | null
          applicant_requirement_id?: number | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profiles_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "applicant_requirement_media_applicant_requirement_id_fkey"
            columns: ["applicant_requirement_id"]
            isOneToOne: false
            referencedRelation: "applicant_requirements"
            referencedColumns: ["id"]
          },
        ]
      }
      applicant_requirements: {
        Row: {
          created_at: string
          id: number
          profile_id: string | null
          remarks: string | null
          requirement_id: number | null
          status: Database["public"]["Enums"]["status_type"] | null
        }
        Insert: {
          created_at?: string
          id?: never
          profile_id?: string | null
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
        }
        Update: {
          created_at?: string
          id?: never
          profile_id?: string | null
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
  public: {
    Tables: {
// ...
      requirement_templates: {
        Row: {
          created_at: string
          description: string | null
          id: number
          name: string | null
          requirement_type: string | null
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: number
          name?: string | null
          requirement_type?: string | null
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: number
          name?: string | null
          requirement_type?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
// ...
  }
// ...
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never